refactor: reorganize search package into subpackages (regex, invindex) and rename searchers for clarity
