+++ internal/common/search/invindex/item_info_set_test.go
@@ -1,4 +1,4 @@
-package search
+package invindex
 
 import (
 	"reflect"
